Refactored with more consistent naming and placement. Fixed the table style wrapper.
