+++ evennia/__init__.py
@@ -110,6 +110,10 @@ TICKER_HANDLER = None
 MONITOR_HANDLER = None
 CHANNEL_HANDLER = None
 
+# Containers
+VALIDATOR_CONTAINER = None
+OPTION_CONTAINER = None
+
 
 
 def _create_version():
@@ -154,6 +158,7 @@ def _init():
     global create_object, create_script, create_account, create_channel, create_message, create_help_entry
     global settings, lockfuncs, logger, utils, gametime, ansi, spawn, managers
     global contrib, TICKER_HANDLER, MONITOR_HANDLER, SESSION_HANDLER, CHANNEL_HANDLER, TASK_HANDLER
+    global VALIDATOR_CONTAINER, OPTION_CONTAINER
     global EvMenu, EvTable, EvForm, EvMore, EvEditor
     global ANSIString 
 
@@ -214,6 +219,10 @@ def _init():
     from .server.sessionhandler import SESSION_HANDLER
     from .comms.channelhandler import CHANNEL_HANDLER
     from .scripts.monitorhandler import MONITOR_HANDLER
+
+    # containers
+    from .utils.containers import VALIDATOR_CONTAINER
+    from .utils.containers import OPTION_CONTAINER
 
     # initialize the doc string
     global __doc__
